requirements.txt
langfuse>=3.0.0
google-cloud-aiplatform>=1.0.0
vertexai>=1.0.0
python-dotenv>=1.0.0
